requirements.txt
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.2
python-dotenv==1.0.0
pypdf==3.17.1
chromadb==0.4.22
openai==1.55.3
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==1.10.16
reportlab==4.0.8